update logout alert
